backend/requirement.txt
# requirement.txt

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
supabase==2.0.3

# ML/AI
sentence-transformers==2.3.1
onnxruntime==1.19.2
transformers==4.34.0  # ← Downgrade to 4.34.0 (stable với torch 2.1.1)
numpy==1.26.4 

# Vietnamese NLP / OCR
vietocr==0.3.13

# Media Processing
yt-dlp>=2024.12.3
opencv-python==4.8.1.78
easyocr==1.7.0
moviepy==1.0.3
imageio-ffmpeg==0.4.9

# Audio Processing
openai-whisper==20231117
librosa==0.10.1
soundfile==0.12.1

# HTTP Client
httpx==0.24.1

# Utils
pydantic==2.5.0

